refactor(emmet): create testing utility for aggregate stream
