+++ src/packages/emmett/src/eventStore/inMemoryEventStore.unit.spec.ts
@@ -1,80 +1,11 @@
 /* eslint-disable @typescript-eslint/no-floating-promises */
-import assert from 'node:assert';
-import { randomUUID } from 'node:crypto';
-import { describe, it } from 'node:test';
+import { describe } from 'node:test';
 import { getInMemoryEventStore } from '../eventStore';
-import {
-  evolve,
-  getInitialState,
-  type PricedProductItem,
-  type ShoppingCartEvent,
-} from '../testing/shoppingCart.domain';
+import { testAggregateStream } from '../testing';
 
 // Events & Entity
 
 describe('InMemoryEventStore', () => {
   const eventStore = getInMemoryEventStore();
-
-  describe('aggregateStream', () => {
-    it('When called with `to` allows time traveling', async () => {
-      // Given
-      const productItem: PricedProductItem = {
-        productId: '123',
-        quantity: 10,
-        price: 3,
-      };
-      const discount = 10;
-      const shoppingCartId = randomUUID();
-
-      await eventStore.appendToStream<ShoppingCartEvent>(shoppingCartId, [
-        { type: 'ProductItemAdded', data: { productItem } },
-      ]);
-      await eventStore.appendToStream<ShoppingCartEvent>(shoppingCartId, [
-        { type: 'ProductItemAdded', data: { productItem } },
-      ]);
-      await eventStore.appendToStream<ShoppingCartEvent>(shoppingCartId, [
-        { type: 'DiscountApplied', data: { percent: discount } },
-      ]);
-
-      // when
-      const resultAt1 = await eventStore.aggregateStream(shoppingCartId, {
-        evolve,
-        getInitialState,
-        read: { to: 1n },
-      });
-      const resultAt2 = await eventStore.aggregateStream(shoppingCartId, {
-        evolve,
-        getInitialState,
-        read: { to: 2n },
-      });
-      const resultAt3 = await eventStore.aggregateStream(shoppingCartId, {
-        evolve,
-        getInitialState,
-        read: { to: 3n },
-      });
-
-      // then
-      assert.ok(resultAt1);
-      assert.ok(resultAt2);
-      assert.ok(resultAt3);
-
-      assert.equal(resultAt1.currentStreamVersion, 1);
-      assert.deepEqual(resultAt1.state, {
-        productItems: [productItem],
-        totalAmount: 30,
-      });
-
-      assert.equal(resultAt2.currentStreamVersion, 2);
-      assert.deepEqual(resultAt2.state, {
-        productItems: [productItem, productItem],
-        totalAmount: 60,
-      });
-
-      assert.equal(resultAt3.currentStreamVersion, 3);
-      assert.deepEqual(resultAt3.state, {
-        productItems: [productItem, productItem],
-        totalAmount: 54,
-      });
-    });
-  });
+  testAggregateStream(eventStore);
 });
