+++ src/packages/emmett/src/testing/index.ts
@@ -1,3 +1,5 @@
 export * from './assertions';
 export * from './deciderSpecification';
+export * from './features';
 export * from './shoppingCart.domain';
+
